apps/server/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metadatahub-server"
version = "0.1.0"
description = "MetadataHub FastAPI backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.40.0",
    "openpyxl>=3.1.0",
    "duckdb>=1.1.0",
    "redis[hiredis]>=5.2.0",
    "python-multipart>=0.0.12",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.0.0",
    "httpx>=0.28.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
